Move fireservicerota coordinator to own module
